requirements.txt
torch
torchvision
opencv-python
Pillow
matplotlib
numpy
tqdm
kagglehub
